commit at 2025.10.5—Change Debug dsiplay to 10samples shot a refresh; this make delaye much less:本次更新内容较多，main_debug初步结构化；debug图表数据更新改为10个sample一次，大幅度降低程序的延迟，基本可以实时检查数据—目前已知的问题：debug数据和displacement数据是分别计算的；debug数据无法支持调试displacement

+++ main_distance_structure/functions_for_structure/calculateDisplacement.m
@@ -4,9 +4,9 @@ function [dx, dy, dz] = calculateDisplacement(accelX, accelY, accelZ, gyroX, gyr
     G = 9.80665;
     
     % 将加速度数据转换为G单位，然后转换为m/s²
-    accelX_g = double((accelX / 32768)) * accel_range;
-    accelY_g = double((accelY / 32768)) * accel_range;
-    accelZ_g = double((accelZ / 32768)) * accel_range;
+    accelX_g = (accelX / 32768) * accel_range;
+    accelY_g = (accelY / 32768) * accel_range;
+    accelZ_g = (accelZ / 32768) * accel_range;
     
     accelX_mps2 = accelX_g * G;
     accelY_mps2 = accelY_g * G;
